pyproject.toml
[tool.poetry]
name = "kokebok-api"
version = "0.1.0"
description = ""
authors = ["Jonatan", "Mia"]
readme = "README.md"
packages = [{include = "kokebok"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
django-ninja = "^0.22.2"
django-environ = "^0.11.2"
psycopg2 = "^2.9.8"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
recipe-scrapers = "^14.49.3"
pillow = "^10.0.1"
django-cors-headers = "^4.2.0"
django-storages = "^1.14.2"
boto3 = "^1.28.62"
django-csp = "^3.7"
google-cloud-vision = "^3.4.4"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
isort = "^5.12.0"
black = {version = "^23.9.1", allow-prereleases = true}
ipython = "^8.15.0" # Improves Django shell experience ("python manage.py shell -i ipython")
types-requests = "^2.31.0.7"
types-beautifulsoup4 = "^4.12.0.6"
parameterized = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


## TOOLING:

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 86

[tool.black]
line-length = 86
target-version = ['py310']
include = '\.pyi?$'
force-exclude = '''
(
  /(
      \.mypy_cache   # exclude a few common directories in the
    | \.git          # root of the project
    | \.venv
    | \.vscode
    |
  )/
  | .*\/migrations\/.*
)
'''

[tool.mypy]
exclude = [
    '.*\/migrations\/.*',
    'kokebok/manage\.py',
    # 'kokebok/kokebok/settings\.py',
]

[[tool.mypy.overrides]]
# Ignore import errors for django and django-ninja
module = [
  'django.*',
  'ninja.*',
  'extruct.*',
  'environ.*',
]
ignore_missing_imports = true
